+++ sdk/cognitiveservices/cognitiveservices-face/src/operations/faceListOperations.ts
@@ -52,25 +52,29 @@ export class FaceListOperations {
    * Please Refer to [Specify a face recognition
    * model](https://docs.microsoft.com/azure/cognitive-services/face/face-api-how-to-topics/specify-recognition-model).
    * @param faceListId Id referencing a particular face list.
+   * @param name User defined name, maximum length is 128.
    * @param [options] The optional parameters
    * @returns Promise<msRest.RestResponse>
    */
-  create(faceListId: string, options?: Models.FaceListCreateOptionalParams): Promise<msRest.RestResponse>;
+  create(faceListId: string, name: string, options?: Models.FaceListCreateOptionalParams): Promise<msRest.RestResponse>;
   /**
    * @param faceListId Id referencing a particular face list.
+   * @param name User defined name, maximum length is 128.
    * @param callback The callback
    */
-  create(faceListId: string, callback: msRest.ServiceCallback<void>): void;
+  create(faceListId: string, name: string, callback: msRest.ServiceCallback<void>): void;
   /**
    * @param faceListId Id referencing a particular face list.
+   * @param name User defined name, maximum length is 128.
    * @param options The optional parameters
    * @param callback The callback
    */
-  create(faceListId: string, options: Models.FaceListCreateOptionalParams, callback: msRest.ServiceCallback<void>): void;
-  create(faceListId: string, options?: Models.FaceListCreateOptionalParams | msRest.ServiceCallback<void>, callback?: msRest.ServiceCallback<void>): Promise<msRest.RestResponse> {
+  create(faceListId: string, name: string, options: Models.FaceListCreateOptionalParams, callback: msRest.ServiceCallback<void>): void;
+  create(faceListId: string, name: string, options?: Models.FaceListCreateOptionalParams | msRest.ServiceCallback<void>, callback?: msRest.ServiceCallback<void>): Promise<msRest.RestResponse> {
     return this.client.sendOperationRequest(
       {
         faceListId,
+        name,
         options
       },
       createOperationSpec,
@@ -349,10 +353,7 @@ const createOperationSpec: msRest.OperationSpec = {
   ],
   requestBody: {
     parameterPath: {
-      name: [
-        "options",
-        "name"
-      ],
+      name: "name",
       userData: [
         "options",
         "userData"